requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langfuse
python-dotenv